requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.7.1
Django==5.0
django-allauth==0.61.1
django-ckeditor==6.7.0
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-sslserver==0.22
idna==3.6
oauthlib==3.2.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
sslcommerz-lib==1.0
tzdata==2023.4
urllib3==2.2.0
